backend/requirements.txt
# Core Agent Framework
smolagents==1.24.0
litellm

# Web Framework
fastapi
uvicorn[standard]
websockets
python-multipart

# Data & Storage
sqlalchemy
aiosqlite

# Utilities
pydantic
python-dotenv
